move all specific connections into test folder
