--- src/fdi.js
@@ -1,10 +1,0 @@
-import { DirectedConnection } from './connection.js'
-
-export class FDI extends DirectedConnection{
-	constructor(source,target){
-		super(source,target)
-	}
-	get id(){
-		return `${this.constructor.name}:${super.id}`
-	}
-}
