+++ src/index.js
@@ -2,7 +2,7 @@ export { JurisdictionGraph } from './graph.js'
 export { Jurisdiction } from './jurisdiction.js'
 export { assignBoundaries } from './fetchGeoms'
 export { Connection, DirectedConnection } from './connection.js'
-export { FDI } from './fdi.js'
-export { Twinning } from './twinning.js'
-export { Mission } from './mission.js'
-export { Business } from './business.js'
+//export { FDI } from './fdi.js'
+//export { Twinning } from './twinning.js'
+//export { Mission } from './mission.js'
+//export { Business } from './business.js'
