+++ tests/connections/mission.js
@@ -1,4 +1,4 @@
-import { DirectedConnection } from './connection.js'
+import { DirectedConnection } from '../../src'
 
 export class Mission extends DirectedConnection {
 	#data
